==> 0-CleanEventsData.ipynb <==
### Big Data - Group Project 
#### Cleaning data


from pyspark.sql import SparkSession
spark = SparkSession.builder.appName('SparkBasics').getOrCreate()
spark

#change configuration settings on Spark 
# change the value of the execute memory . Same as at work. 
conf = spark.sparkContext._conf.setAll([('spark.executor.memory', '4g'),
                                        ('spark.app.name', 'Spark Updated Conf'), 
                                        ('spark.executor.cores', '4'), 
                                        ('spark.cores.max', '4'), 
                                        ('spark.driver.memory','4g')])

sc = spark.sparkContext
sc.defaultParallelism
sc._jsc.sc().getExecutorMemoryStatus().size()

#Check data exists in location
!hadoop fs -ls -d "<hdfs location>"

# this is the right way to read a file - from HDFS in this case
%time df = spark.read.csv("/user/josefinabollini/GroupProject/games_CUBS_socks.csv", inferSchema=True, header=True)
import pyspark.sql.functions as f
df = df.withColumn('MONTH_ok',f.col('DAY'))
df = df.withColumn('DAY_ok',f.col('MONTH')).drop('MONTH','DAY')
df =  df.withColumnRenamed('MONTH_ok','MONTH').withColumnRenamed('DAY_ok','DAY')
df = df.withColumn('NEIGHBORHOOD',f.when(f.col('LOCATION') == 'Wrigley Field','Lakeview').otherwise('Armour Square'))
df = df.withColumn('nb_code',f.when(f.col('LOCATION') == 'Wrigley Field',6).otherwise(34))
df.show(5)

%time other_events_df = spark.read.csv("/user/josefinabollini/GroupProject/otherEvents.csv", inferSchema=True, header=True)
other_events_df =  other_events_df.withColumnRenamed('Location','LOCATION').withColumnRenamed('NEIGHBOURHOOD','NEIGHBORHOOD')
other_events_df.show(5)

df = df.select("LOCATION","DAY","MONTH","YEAR","NEIGHBORHOOD",'TEAM','nb_code')
other_events_df = other_events_df.select("LOCATION","DAY","MONTH","YEAR","NEIGHBORHOOD",'TEAM','nb_code')

df3 = df.unionAll(other_events_df)
df3.describe().show()

df3.toPandas().to_csv("/home/josefinabollini/GroupProject/All_events_aligned.csv")

### Weather Data - Cleaning


%time df21 = spark.read.csv("/user/josefinabollini/GroupProject/2021weatherdata.csv", inferSchema=True, header=True)
%time df22 = spark.read.csv("/user/josefinabollini/GroupProject/2022weatherdata.csv", inferSchema=True, header=True)
%time df1920 = spark.read.csv("/user/josefinabollini/GroupProject/2019-2020weatherdata.csv", inferSchema=True, header=True)



df1920.show(10)
