==> 1-DataEngineering-Compression.ipynb <==
# Data Engineering

## Turn all files into parquet for compression




#create Spark session
spark = SparkSession.builder.appName('turntoparquet').getOrCreate()

#turn chicago communities csv into parquet

chi_communities = spark.read \
    .option("quote", "\"")  \
    .option("escape", "\"") \
    .option("ignoreLeadingWhiteSpace",True) \
    .option("multiline", True)\
    .csv("gs://big-data-final/raw-chicago-communities/chi_communities.csv",inferSchema=True, header=True )

chi_communities.show()

chi_communities.write.parquet('gs://big-data-final/clean-parquet/chi_communities.parquet')

par_df = spark.read.parquet("gs://big-data-final/clean-parquet/chi_communities.parquet")

par_df.show()

#turn chicago communities csv into parquet

covid = spark.read \
    .option("quote", "\"")  \
    .option("escape", "\"") \
    .option("ignoreLeadingWhiteSpace",True) \
    .option("multiline", True)\
    .csv("gs://big-data-final/raw-covid/covid.csv",inferSchema=True, header=True )

covid.show()

from pyspark.sql import functions as F

renamed_df = covid.select([F.col(col).alias(col.replace(' ', '_')) for col in covid.columns])

renamed_df.show()

renamed_df.write.parquet('gs://big-data-final/clean-parquet/covid.parquet')

#turn events csv into parquet

events = spark.read \
    .option("quote", "\"")  \
    .option("escape", "\"") \
    .option("ignoreLeadingWhiteSpace",True) \
    .option("multiline", True)\
    .csv("gs://big-data-final/raw-events/All_events_aligned.csv",inferSchema=True, header=True )

events.show()

events.write.parquet('gs://big-data-final/clean-parquet/events.parquet')

#turn weather csv into parquet

w20192020 = spark.read \
    .option("quote", "\"")  \
    .option("escape", "\"") \
    .option("ignoreLeadingWhiteSpace",True) \
    .option("multiline", True)\
    .csv("gs://big-data-final/raw-weather/2019-2020 weather data.csv",inferSchema=True, header=True )

w20192020.show()

w20192020.write.parquet('gs://big-data-final/clean-parquet/w20192020.parquet')

w2021 = spark.read \
    .option("quote", "\"")  \
    .option("escape", "\"") \
    .option("ignoreLeadingWhiteSpace",True) \
    .option("multiline", True)\
    .csv("gs://big-data-final/raw-weather/2021 weather data.csv",inferSchema=True, header=True )

w2021.show()

w2021.write.parquet('gs://big-data-final/clean-parquet/w2021.parquet')

rideshare = spark.read \
    .option("quote", "\"")  \
    .option("escape", "\"") \
    .option("ignoreLeadingWhiteSpace",True) \
    .option("multiline", True)\
    .csv("gs://big-data-final/raw-rideshare/rideshare.csv",inferSchema=True, header=True )

rideshare.show(3)

from pyspark.sql import functions as F

rideshare_df = rideshare.select([F.col(col).alias(col.replace(' ', '_')) for col in rideshare.columns])

rideshare_df.show(3)

rideshare_df.write.parquet('gs://big-data-final/clean-parquet/rideshare.parquet')
print('finished compression')

taxi = spark.read \
    .option("quote", "\"")  \
    .option("escape", "\"") \
    .option("ignoreLeadingWhiteSpace",True) \
    .option("multiline", True)\
    .csv("gs://big-data-final/raw-taxi/taxi.csv",inferSchema=True, header=True )

taxi.show(3)

taxi_df = taxi.select([F.col(col).alias(col.replace(' ', '_')) for col in taxi.columns])

taxi_df.show(3)

taxi_df.write.parquet('gs://big-data-final/clean-parquet/taxi.parquet')
print('finished compression')
